sharing: harden ShareRanking code
In particular:
* Many DCHECKs are promoted into DCHECKs
* Indexes and lengths are represented with size_t instead of unsigned
  int
* (begin, end) std::vector constructions are replaced with uses of
  base::ToVector(span)

There is at least one memory corruption bug lurking in this code (the
linked one) so hopefully this will help flush it out.

Bug: 331577262
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5698134
Commit-Queue: Elly FJ <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1328321}

+++ chrome/browser/share/share_ranking.h
@@ -48,8 +48,8 @@ class ShareRanking : public base::SupportsUserData::Data {
   void Rank(ShareHistory* history,
             const std::string& type,
             const std::vector<std::string>& available_on_system,
-            unsigned int fold,
-            unsigned int length,
+            size_t fold,
+            size_t length,
             bool persist_update,
             GetRankingCallback callback);
 
@@ -68,8 +68,8 @@ class ShareRanking : public base::SupportsUserData::Data {
       const std::map<std::string, int>& recent_share_history,
       const Ranking& old_ranking,
       const std::vector<std::string>& available_on_system,
-      unsigned int fold,
-      unsigned int length,
+      size_t fold,
+      size_t length,
       Ranking* display_ranking,
       Ranking* persisted_ranking);
 
@@ -90,8 +90,8 @@ class ShareRanking : public base::SupportsUserData::Data {
     // These members mirror the parameters passed in to Rank().
     std::string type;
     std::vector<std::string> available_on_system;
-    unsigned int fold;
-    unsigned int length;
+    size_t fold;
+    size_t length;
     bool persist_update;
     GetRankingCallback callback;
